:sparkles: Containerize omnisharp-roslyn
[Omnisharp-Roslyn](https://github.com/OmniSharp/omnisharp-roslyn)
is a language server protocol implementation that provides C#
language services, such as code completion, code navigation, and
diagnostics, enabling enhanced development experiences in various text
editors and IDEs. It leverages the Roslyn compiler platform to analyze
C# code and provide real-time feedback to developers for efficient code
editing and troubleshooting.

This puts the stdio server in a container for use as a baseline as we
work to include it as an external provider for analyzer-lsp.

Signed-off-by: David Zager <david.j.zager@gmail.com>